# dc-api/openapi.yaml
openapi: 3.0.3
info:
  title: DC-API
  version: 0.1.0
  description: |
    WSO2 Sovereign Cloud Control Plane API.

    DC-API sits in front of a Rancher + Harvester infrastructure layer and
    exposes a clean, tenant-isolated REST API for managing VMs, RKE2 clusters,
    virtual networks, DNS zones, and RBAC — in the style of AWS / Azure.

    **Authentication**

    Every `/v1/*` endpoint requires a `Bearer` token in the `Authorization`
    header. Two token types are accepted by the same scheme:

    1. **Asgardeo OIDC JWT** — obtained via the `dcctl login` PKCE flow. The JWT
       must carry a `groups` claim containing at least one group matching the
       tenant group prefix (default `dc-tenant-<tenant-id>`).

    2. **Service-account token** — a long-lived opaque token with the format
       `dcapi_sa_<lookup_id>_<secret>`, issued by
       `POST /v1/tenants/{tenant_id}/service-accounts`. Store it on creation;
       it cannot be retrieved again.

    **Async provisioning**

    Long-running operations (VM create, VNet create, etc.) return `202 Accepted`
    immediately. The response body contains the resource object with
    `"status": "PENDING"` and a `note` field with the polling URL. Poll the
    GET endpoint until `status` transitions to `ACTIVE` or `FAILED`.

    **Status lifecycle**

    `PENDING → ACTIVE` or `PENDING → FAILED`

    Deletion: `ACTIVE → DELETING → (row removed on success)`

    **Error format**

...
servers:
...
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /v1/tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants the authenticated principal has access to
      description: |
        Returns the set of tenants for which the authenticated principal holds
        at least one `role_assignments` row (at any scope within the tenant).

        The `roles` array is the **union of distinct roles** the principal holds
        across all scopes within that tenant — useful for the UI tenant switcher
        to decide which actions to enable.

        Both human users (OIDC JWT) and service accounts (dcapi_sa_* token)
        may call this endpoint. The result is always scoped to the caller's
        own principal.

        **Platform admins** (sub in `DCAPI_PLATFORM_ADMIN_SUBS`, OR member of
        `DCAPI_ADMIN_GROUP`): this endpoint returns **every tenant** in the
        `tenants` registry — not just the caller's role-assignment rows.
        Each entry's `roles` is set to `["owner"]` for UI convenience;
        admins bypass per-tenant RBAC checks at the middleware layer
        regardless. Tenants without role-assignment rows are still
        visible to admins as long as they're in the registry — the
        registry is populated by `POST /v1/admin/tenants` (explicit
        admin pre-registration) or by `POST /v1/tenants/{tid}/members`
        (UPSERT on first invite).

        **Required auth**: any valid token. No minimum role required beyond
        being authenticated.
      responses:
        "200":
          description: List of tenants the principal has access to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantSummary"
              example:
                - id: iam-platform
                  name: iam-platform
                  roles:
                    - owner
                - id: dev-team
                  name: dev-team
                  roles:
                    - member
                    - viewer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Admin: tenant registry ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Accepts either:
        1. An Asgardeo OIDC JWT (RS256) — obtained via the PKCE flow (`dcctl login`).
        2. A DC-API service-account token with the format
           `dcapi_sa_<lookup_id>_<secret>` — issued by
           `POST /v1/tenants/{tenant_id}/service-accounts`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "missing Authorization header"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "failed to register resource"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong
          example: "name is required"
...
    TenantSummary:
      type: object
      required: [id, name, roles]
      description: |
        A tenant the authenticated principal has access to, along with
        the union of roles the principal holds across all scopes within
        that tenant.
      properties:
        id:
          type: string
          description: Tenant slug identifier
          example: iam-platform
        name:
          type: string
          description: Display name (same as id in M1.5 — separate name field is M5)
          example: iam-platform
        roles:
          type: array
          items:
            type: string
            enum: [owner, member, viewer]
          description: |
            Union of distinct roles the principal holds across all scopes
            within this tenant. Used by the UI to decide which actions to enable.
          example:
            - owner
...